AuthJwt/requirement.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
psycopg2==2.8.6